% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ajoute_coursdeau.R
\name{ajoute_coursdeau}
\alias{ajoute_coursdeau}
\title{ajoute_coursdeau}
\usage{
ajoute_coursdeau(
  shp,
  con,
  schema = "r400_cours_deau",
  table = "ice_vilaine",
  color = "blue",
  size = 0.8
)
}
\arguments{
\item{shp}{Objet sf utilisé pour filtrer spatialement}

\item{con}{Connexion DBI à la base PostGIS}

\item{schema}{Nom du schéma PostGIS}

\item{table}{Nom de la table PostGIS}

\item{color}{Couleur du tracé}

\item{size}{Épaisseur du tracé}
}
\value{
Un layer ggplot2::geom_sf
}
\description{
Ajoute une couche de cours d'eau filtrée spatialement à un ggplot
}
\examples{


if (interactive()) {
  config_path <- "C://workspace//gwilenalim//yaml//config.yml"
  if (file.exists(config_path)) {
    library(RPostgres)
    config <- yaml::read_yaml(config_path)
    con <- tryCatch({
      DBI::dbConnect(
        RPostgres::Postgres(),
        host = config$host,
        port = config$port,
        user = config$user,
        password = config$password,
        dbname = config$dbname
      )
    }, error = function(e) NULL)
  
    if (!is.null(con)) {
       triangle_sf <- sf::st_sf(
        geometry = sf::st_sfc(
          sf::st_polygon(list(rbind(
            c(-1.6794, 48.1147),  # Rennes
            c(-1.2100, 48.1230),  # Vitré
            c(-1.5025, 47.6833),  # Derval
            c(-1.6794, 48.1147)   # Retour à Rennes
          )))
        ),
        crs = 4326
      )
     
     g<- f_fond_de_carte(triangle_sf, zoom = 9) +
    ajoute_coursdeau(triangle_sf, con, size = 0.5)

      DBI::dbDisconnect(con)
      print(g)
    } else {
      message("Connexion à la base impossible, exemple non exécuté.")
    }
  } else {
    message("Fichier de configuration introuvable.")
  }
}

knitr::include_graphics(system.file("extdata", "graph_coursdeaubv.png", package = "valorisationQE"))

}
