% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_fonctions_pesticides.R
\name{analyse_fonctions_pesticides}
\alias{analyse_fonctions_pesticides}
\title{analyse_fonctions_pesticides}
\usage{
analyse_fonctions_pesticides(
  data,
  code_remarque = "CdRqAna",
  code_parametre = "CdParametre",
  liste_pesticides
)
}
\arguments{
\item{data}{Data frame contenant les résultats analytiques, incluant les colonnes
correspondant aux codes de remarque et de paramètre.}

\item{code_remarque}{Nom de la colonne dans \code{data} correspondant au code de remarque analytique.
Par défaut \code{"CdRqAna"}.}

\item{code_parametre}{Nom de la colonne dans \code{data} correspondant au code du paramètre.
Par défaut \code{"CdParametre"}.}

\item{liste_pesticides}{Data frame issu de la fonction \code{importe_ref_pestibase}, contenant
les colonnes \code{SA_CodeSANDRE}, \code{function_bzh} et \code{function_fr} pour l'association des fonctions.}
}
\value{
Un objet \code{ggplot} représentant la répartition des fonctions des substances
pesticides quantifiées.
}
\description{
Analyse des fonctions des substances pesticides
}
\details{
Cette fonction permet d'analyser les fonctions des substances pesticides quantifiées
à partir d'un jeu de données analytique et d'une table de référence des pesticides.
Elle filtre les résultats selon un code de remarque analytique, joint les données
aux fonctions des substances, puis génère un graphique ggplot représentant la
répartition des fonctions en pourcentage de la concentration totale mesurée.
}
\examples{

ref_pestibase <- importe_ref_pestibase()

data <- data.frame(
   DatePrel = Sys.Date() + rep(sort(sample(1:500, 10)), 3),
   RsAna = round(runif(60, 0, 0.5), 2),
   LqAna = 0.1,
   CdStationMesureEauxSurface = c("A", "B", "C"),
   CdParametre = sample(ref_pestibase$liste_phytos_autorises$SA_CodeSANDRE,
                        size=10,
                        replace=TRUE),
   CdUniteMesure = "133"
 )
 data$CdRqAna <- ifelse(data$RsAna >= data$LqAna, "1", "10")

analyse_fonctions_pesticides(data, 
                             code_remarque="CdRqAna",
                             code_parametre="CdParametre",
                             liste_pesticides=ref_pestibase$liste_phytos_autorises)


}
