% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_cartographie_concentrations.R
\name{f_cartographie_concentrations}
\alias{f_cartographie_concentrations}
\title{f_cartographie_concentrations}
\usage{
f_cartographie_concentrations(
  fond_carte,
  shp_staq,
  data_carte,
  col_stations = "CdStationMesureEauxSurface",
  col_valeurs = "RsAna",
  titre = "Bilan des concentrations en pesticides",
  sous_titre = "Date de la campagne",
  breaks = c(0, 0.5, 2, Inf),
  values = c("cyan", "yellow", "red", "black"),
  pts_size = 6,
  text_size = 5,
  unite = "µg/L",
  nom_legende = "Concentrations",
  nb_decimal_text = 1,
  label.size = 0.25,
  box.padding = 0.35,
  point.padding = 0.3,
  segment.color = "grey50",
  min.segment.length = 0
)
}
\arguments{
\item{fond_carte}{Objet \code{ggplot} représentant la carte de fond, par exemple
généré avec la fonction f_fond_de_carte}

\item{shp_staq}{Objet sf contenant les points de suivis à afficher. Seuls les
points dont le code SANDRE est présent dans data_carte seront affichés.}

\item{data_carte}{Data frame contenant les données de prélèvement, incluant
les colonnes de date de prélèvement, de code SANDRE des stations et la
colonne de valeurs à cartographier.}

\item{col_stations}{Nom de la colonne contenant les codes sandre des stations
de mesure à afficher (par défaut \code{"CdStationMesureEauxSurface"}). Si ce nom
de colonne est différent pour shp_staq et data_carte alors col_station
est un vecteur dont la première valeur est le nom de la colonne de shp_staq
et la deuxième celle de data_carte}

\item{col_valeurs}{Nom de la colonne contenant les valeurs de concentration à
utiliser (par défaut \code{"RsAna"}).}

\item{titre}{Titre principal de la carte (par défaut \code{"Bilan des concentrations en pesticides"}).}

\item{sous_titre}{Sous-titre de la carte.}

\item{breaks}{Vecteur numérique définissant les seuils de classes de
concentration (par défaut \code{c(0, 0.5, 2, Inf)}).}

\item{values}{Vecteur de couleurs correspondant aux classes de concentration
(par défaut \code{c("cyan", "yellow", "red", "black")}).}

\item{pts_size}{Taille des points représentant les stations (par défaut \code{6}).}

\item{text_size}{Taille du texte des étiquettes (par défaut \code{5}).}

\item{unite}{Unité des concentrations affichée dans la légende et les
étiquettes (par défaut \code{"µg/L"}).}

\item{nom_legende}{Titre de la légende des couleurs (par
défaut \code{"Concentrations"}).}

\item{nb_decimal_text}{Nombre de décimales à afficher dans les étiquettes
(par défaut \code{1}).}

\item{label.size}{taille de l'étiquette (défaut = 0.25). Voir geom_text_repel
du package ggrepel
(par défaut \code{1}).}

\item{box.padding}{Marge autour de l'étiquette (par défaut 0.35) Voir
geom_text_repel du package ggrepel}

\item{point.padding}{Marge autour de l'étiquette (par défaut 0.3) Voir
geom_text_repel du package ggrepel}

\item{segment.color}{Couleur du segment liant les points à leurs étiquettes.
grey50 par défaut. Voir geom_text_repel du package ggrepel}

\item{min.segment.length}{Longueur mini entre un point et son étiquette
(par défaut 0) Voir geom_text_repel du package ggrepel}
}
\value{
Un objet \code{ggplot} représentant la carte des concentrations avec
légende, points colorés selon les classes, et étiquettes.
}
\description{
Génère une carte des concentrations en polluants
}
\details{
Cette fonction permet de représenter spatialement les concentrations en
polluants mesurées lors d'une campagne de prélèvements, en utilisant une
carte de fond et des données géolocalisées. Les concentrations sont classées
selon des seuils définis et affichées avec une légende personnalisée.
}
\examples{

library(sf)
library(dplyr)

# Coordonnées approximatives des centres des communes
coords <- data.frame(
  code = c("A", "B", "C"),
  commune = c("Châteaugiron", "Janzé", "Thourie"),
  lon = c(-1.499, -1.498, -1.4833),
  lat = c(48.052, 47.960, 47.8550)
)

# Création de l'objet sf
points_sf <- st_as_sf(coords, coords = c("lon", "lat"), crs = 4326)


# bassin versant
   triangle_sf <- sf::st_sf(
        geometry = sf::st_sfc(
          sf::st_polygon(list(rbind(
            c(-1.6794, 48.1147),  # Rennes
            c(-1.2100, 48.1230),  # Vitré
            c(-1.5025, 47.6833),  # Derval
            c(-1.6794, 48.1147)   # Retour à Rennes
          )))
        ),
        crs = 4326
      )

fond_carte<-f_fond_de_carte(triangle_sf, zoom=10)

# Création du data.frame
df_valeurs <- data.frame(
  code = c("A", "B", "C"),
  valeur = c(0.23, 0.71, 3.52)
)

f_cartographie_concentrations(fond_carte,
                              shp_staq=points_sf,
                              data_carte=df_valeurs,
                              col_stations="code",
                              col_valeurs="valeur",
                              titre="Bilan des concentrations en pesticides",
                              sous_titre="Date de la campagne",
                              breaks=c(0, 0.5, 2, Inf),
                              values = c("cyan", "yellow", "red", "black"),
                              pts_size=6,
                              text_size=5,
                              unite="µg/L",
                              nom_legende="Concentrations",
                              nb_decimal_text=1)

}
