% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_fond_de_carte.R
\name{f_fond_de_carte}
\alias{f_fond_de_carte}
\title{f_fond_de_carte}
\usage{
f_fond_de_carte(shp, zoom = NULL)
}
\arguments{
\item{shp}{= objet sf}

\item{zoom}{= niveau de zoom openstreetmap (paramètre facultatif)}
}
\value{
un ggplot avec le fond de carte openstreetmap correspondant à l'emprise de
l'objet en entrée
}
\description{
Fonction pour charger un fon de carte openstreetmap correspondant à l'emprise
d'un objet sf
}
\examples{

   triangle_sf <- sf::st_sf(
        geometry = sf::st_sfc(
          sf::st_polygon(list(rbind(
            c(-1.6794, 48.1147),  # Rennes
            c(-1.2100, 48.1230),  # Vitré
            c(-1.5025, 47.6833),  # Derval
            c(-1.6794, 48.1147)   # Retour à Rennes
          )))
        ),
        crs = 4326
      )

f_fond_de_carte(triangle_sf, zoom=9)
}
