% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_substances.R
\name{heatmap_substances}
\alias{heatmap_substances}
\title{heatmap_substances}
\usage{
heatmap_substances(
  df,
  col_station = "CdStationMesureEauxSurface",
  col_param = "NomParametre",
  col_result = "RsAna",
  col_remarque = "CdRqAna",
  titre = "Substances quantifiées au-delà du seuil",
  seuil_bas = 0.1,
  seuil_haut = 2,
  chiffres_significatifs = 2
)
}
\arguments{
\item{df}{Data frame contenant les résultats d’analyses.}

\item{col_station}{Nom de la colonne contenant les codes des stations.
Défaut : \code{"CdStationMesureEauxSurface"}.}

\item{col_param}{Nom de la colonne contenant les noms des substances ou
paramètres. Défaut : \code{"NomParametre"}.}

\item{col_result}{Nom de la colonne contenant les résultats d’analyse.
Défaut : \code{"RsAna"}.}

\item{col_remarque}{Nom de la colonne contenant les codes de remarque.
Défaut : \code{"CdRqAna"}.}

\item{titre}{Titre du graphique. Défaut : \code{"Substances quantifiées au-delà  du seuil"}.}

\item{seuil_bas}{Seuil inférieur du gradient de couleur (début du jaune).
Défaut : \code{0.1}.}

\item{seuil_haut}{Seuil supérieur du gradient de couleur (début du rouge).
Défaut : \code{2}.}

\item{chiffres_significatifs}{Nombre de chiffres significatifs à afficher
dans les cases. Défaut : \code{2}.}
}
\value{
Un objet ggplot représentant la heatmap.
}
\description{
Génère une heatmap des substances quantifiées au-delà d’un seuil
}
\details{
Cette fonction crée une heatmap des concentrations de substances (ex. :
pesticides)
mesurées dans différentes stations, en mettant en évidence celles dont les
valeurs dépassent un seuil donné.
Le gradient de couleur est personnalisable selon les seuils définis.
}
\examples{


# Définition des stations
stations <- c("04123456", "04345678", "04987654", "04246810")

# Définition des pesticides courants en France
pesticides <- c(
  "Glyphosate", "Atrazine", "Métolachlore", "S-métolachlore", "Terbuthylazine",
  "Chlorotoluron", "Isoproturon", "Diuron", "Pendiméthaline", "Prosulfocarbe",
  "Acétochlore", "Alachlore", "Linuron", "Oxadiazon", "Clopyralid"
)

# Génération du jeu de données
set.seed(123)  # Pour reproductibilité
df_fictif <- expand.grid(CdStationMesureEauxSurface = stations, NomParametre = pesticides) \%>\%
  dplyr::rowwise() \%>\%
  dplyr::mutate(
    RsAna = round(runif(1, 0, 1), 2),
    CdRqAna = ifelse(RsAna <= 0.06, "0", "1")
  ) \%>\%
  dplyr::ungroup()


heatmap_substances(
  df = df_fictif,
  titre = "Substances quantifiées > 0.1 µg/L",
  seuil_bas = 0.1,
  seuil_haut = 2,
  chiffres_significatifs=2
)
}
