% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentile_dce.R
\name{percentile_dce}
\alias{percentile_dce}
\title{Calcul du percentile DCE selon la méthode de Hazen}
\usage{
percentile_dce(A, type = 0.9, na.rm = TRUE)
}
\arguments{
\item{A}{Vecteur numérique contenant les valeurs mesurées.}

\item{type}{Numérique compris entre 0 et 1 (ex : 0.1 ou 0.9).}

\item{na.rm}{Logique : TRUE pour ignorer les valeurs manquantes.}
}
\value{
Une valeur numérique correspondant au percentile DCE calculé.
}
\description{
Cette fonction calcule un percentile DCE (ex : 0.1 ou 0.9) en appliquant
strictement la méthode de Hazen, conformément à l'arrêté ministériel
du 25 janvier 2010 modifié relatif à l’évaluation de l’état des eaux.
}
\details{
La méthode de Hazen consiste à :
\itemize{
\item trier les valeurs croissantes,
\item calculer le rang : \eqn{R = (n * p) + 0.5},
\item appliquer un arrondi à l'entier le plus proche,
\item renvoyer la valeur correspondant au rang R.
}
}
\examples{
percentile_dce(c(1, 3, 8, 10, 12), type = 0.9)
percentile_dce(c(4, 6, 9, NA, 3), type = 0.1, na.rm = TRUE)

vec <- c(2.1, 5.4, 3.3, 8.9, 7.1, 12.5, 7.0, 8.5, 2, 4.8, 17, 19.2, NA)
percentile_dce(vec, type = 0.9, na.rm = TRUE)
percentile_dce(vec, type = 0.1, na.rm = TRUE)
}
