% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentile_dce_par_station_multi.R
\name{percentile_dce_par_station_multi}
\alias{percentile_dce_par_station_multi}
\title{Calcul du percentile DCE (Hazen) par station pour plusieurs paramètres}
\usage{
percentile_dce_par_station_multi(
  analyses,
  cd_parametres,
  type = 0.9,
  station_col = "CdStationMesureEauxSurface",
  param_col = "CdParametre",
  result_col = "RsAna"
)
}
\arguments{
\item{analyses}{Tibble de résultats analytiques.}

\item{cd_parametres}{Vecteur de codes paramètres SANDRE à filtrer.}

\item{type}{Percentile DCE (0.1 ou 0.9).}

\item{station_col}{Nom de la colonne station (défaut : CdStationMesureEauxSurface).}

\item{param_col}{Nom de la colonne paramètre (défaut : CdParametre).}

\item{result_col}{Nom de la colonne résultat (défaut : RsAna).}
}
\value{
Une tibble contenant :
\itemize{
\item station_col
\item param_col
\item percentile_dce
\item N : nombre de données non NA
}
}
\description{
Cette fonction applique la méthode de Hazen (via \code{percentile_dce()})
pour calculer un percentile DCE (ex : 0.1 ou 0.9) pour chaque combinaison
station × paramètre, dans une table d'analyses au format SANDRE/ROSEAU.
}
\details{
Les noms des colonnes utilisées peuvent être paramétrés, avec valeurs par défaut :
\itemize{
\item station_col = "CdStationMesureEauxSurface"
\item param_col = "CdParametre"
\item result_col = "RsAna"
}

Un contrôle strict des types est effectué :
\itemize{
\item la colonne station doit être de type caractère ou facteur,
\item la colonne paramètre doit être caractère ou facteur,
\item la colonne résultats doit être numérique.
}
}
\examples{
analyses_test <- tibble::tibble(
  CdStationMesureEauxSurface = c("A","A","B","B","B","C"),
  CdParametre = c("1302","1302","1302","8727","8727","1302"),
  RsAna = c(1,5,4,8,12,7)
)

percentile_dce_par_station_multi(
  analyses_test,
  cd_parametres = c("1302","8727"),
  type = 0.9
)



# Exemple d'utilisation avec colonnes par défaut

analyses_test <- tibble::tibble(
  CdStationMesureEauxSurface = c("A","A","B","B","B","C"),
  CdParametre = c("1302","1302","1302","8727","8727","1302"),
  RsAna = c(1,5,4,8,12,7)
)

percentile_dce_par_station_multi(
  analyses_test,
  cd_parametres = c("1302", "8727"),
  type = 0.9
)

# Exemple avec colonnes personnalisées

analyses_custom <- analyses_test |>
  dplyr::rename(
    station = CdStationMesureEauxSurface,
    param = CdParametre,
    resultat = RsAna
  )

percentile_dce_par_station_multi(
  analyses_custom,
  cd_parametres = c("1302", "8727"),
  type = 0.1,
  station_col = "station",
  param_col = "param",
  result_col = "resultat"
)

}
