% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valorisationQE-package.R
\docType{package}
\name{valorisationQE-package}
\alias{valorisationQE}
\alias{valorisationQE-package}
\title{valorisationQE: Package Pour Valoriser Données QE}
\description{
Ce package sert à valoriser graphiquement des données qualité d'eau.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/AnthonyDEBUR/valorisationQE}
  \item Report bugs at \url{https://github.com/AnthonyDEBUR/valorisationQE/issues}
}

}
\author{
\strong{Maintainer}: Anthony DE BURGHRAVE \email{anthony.deburghrave@eaux-et-vilaine.bzh}

}
\keyword{internal}
