% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiche_usages_pesticides.R
\name{fiche_usages_pesticides}
\alias{fiche_usages_pesticides}
\title{fiche_usages_pesticides
Génère un tableau graphique des usages autorisés ou interdits des pesticides par culture}
\usage{
fiche_usages_pesticides(liste_cd_sandre, liste_pesticides)
}
\arguments{
\item{liste_cd_sandre}{Character vector. Liste des codes SANDRE des masses d'eau pour lesquelles on souhaite afficher les usages.}

\item{liste_pesticides}{Data frame. Issu de la fonction \code{importe_ref_pestibase}, contenant au minimum les colonnes :
\itemize{
\item \code{SA_CodeSANDRE} : code SANDRE de la masse d'eau
\item \code{id} : identifiant pestibase de la substance
\item \code{name_fr} : nom de la substance active
\item \code{function_finale} : fonction associée à la substance
\item \code{has_authorized_oepp_culture} : texte décrivant les cultures pour lesquelles la substance est autorisée
\item \code{has_removed_authorized_oepp_culture} : texte décrivant les cultures pour lesquelles l'autorisation a été retirée
\item \code{has_variant} : texte listan les identifiants pestibases de variants
}}
}
\value{
Un objet graphique de type \code{grob} représentant un tableau coloré des usages par culture.
Ce tableau peut être affiché avec \code{grid::grid.draw()}.
}
\description{
Cette fonction permet de visualiser, sous forme de tableau coloré, les usages des substances actives
(pesticides) selon leur statut réglementaire : autorisé, interdit ou jamais autorisé, pour différentes cultures.
Les couleurs indiquent :
\itemize{
\item \strong{vert} : la substance est actuellement autorisée sur la culture
\item \strong{rouge} : la substance a été retirée de l'autorisation
\item \strong{blanc} : la substance n'a jamais été autorisée sur la culture
}
}
\examples{
ref_pestibase <- importe_ref_pestibase()

fiche_usages_pesticides(liste_cd_sandre=c("2009", "1221", "1882", "5817", "5617", "1678"),
                                        liste_pesticides=ref_pestibase$liste_pesticides)


}
